#2 corrigiendo el monto maximo

+++ Models/Metas.cs
@@ -12,7 +12,7 @@ namespace Parcial1_AP1_EliezerTerrero.Models
         public DateTime Fecha { get; set; }
 
         //[Required(ErrorMessage = "El monto es requerido")]
-        [Range(1, int.MaxValue, ErrorMessage ="El monto es requerido")]
+        [Range(1, double.MaxValue, ErrorMessage ="El monto es requerido")]
         public Decimal Monto { get; set; }
     }
 }
